+++ both/views/SupplierRequestEmailView.jsx
@@ -7,42 +7,62 @@ SupplierRequestEmailView = React.createClass({
     mixins: [ReactMeteorData],
 
     getMeteorData() {
-        var query, request, team, supplier, secret, expiry, token;
-        query = this.props.item;
-        request = Issues.findOne(query);
-        token = this.props.token;
-        if(token) {
-            secret = token.token;
-            expiry = token.expiry;
+        var recipient, request, secret, expiry;
+        request = Issues.findOne(this.props.item);
+        recipient = Users.findOne(this.props.recipient);
+        if(this.props.token) {
+            secret = this.props.token.token;
+            expiry = this.props.token.expiry;
         }
         return {
+            recipient:recipient,
+            request:request,
+            team:request.getTeam(),
+            supplier:request.getSupplier(),
+            facility:request.getFacility(),
             secret:secret,
             expiry:expiry,
-            supplier:request.getSupplier(),
-            team:request.getTeam(),
-            facility:request.getFacility(),
-            request:request,
         }
     },
 
     render() {
-        var team = this.data.team;
-        var supplier = this.data.supplier;
+        var recipient = this.data.recipient;    //the recipient (NB:may be a team)
+
+        if(!recipient) {
+            return;
+        }
+
+        console.log(recipient);
+
+        var team = this.data.team;              //the issuing team
+        var supplier = this.data.supplier;      //the supplier item has been issued to
         var request = this.data.request;
         var facility = this.data.facility;
         var secret = this.data.secret;
         var url;
+
+        var recipientName = "";
+
+        if(recipient.profile&&recipient.profile.firstName) {
+            recipientName = " "+recipient.profile.firstName;
+        }
+        else if(recipient.getName) {
+            recipientName = " "+recipient.getName();
+        }
+
         if(secret) {
             url = Meteor.absoluteUrl('u/'+ secret + '/' + request.getEncodedPath());
         }
         else {
             url = request.getUrl();
         }
+
         var dueDate = moment(request.dueDate).format('MMMM Do YYYY, h:mm:ss a');
         var expiry = this.data.expiry?moment(this.data.expiry).fromNow():null;
+
         return(
             <div>
-                <p>Hi {supplier.getName()},</p>
+                <p>Hi{recipientName},</p>
                 <p>{team.getName()} has just issued you a work order <a href={url}>#{request.code} - {request.getName()} at {facility.getName()}</a>.</p>
                 <p>The priority is listed as {request.priority} and it is due to be completed by {dueDate}.</p>
                 <p>To access full details of the work order, please click the above link (no login required).&nbsp;